Step 1 hardening: fix dword-read UB, unsigned Write* signatures, RNG fidelity
Addresses external review of the step-1 commit before step 2 (BallSprite/
PlayerSprite/TeamData) builds on top of this layer:

- swosReadDword() had undefined behaviour: shifting a promoted uint8_t left
  by 24 is not representable as a signed int once the byte is >= 0x80. Fixed
  by casting each byte to uint32_t before shifting (also applied to
  swosReadWord() for one consistently-correct pattern). Added a test byte
  (0xFF000000) that would have caught this.
- swosWriteWord()/swosWriteDword() took plain `int`; converting a value like
  0xDEADBEEF to it, or right-shifting a negative int, are both
  implementation-defined in C. Now take uint16_t/uint32_t; callers needing
  to write a signed value cast explicitly.
- swosRngReseed() was tagged VERIFIED_PC like the rest of Rng.cs, but
  ../swos-port/src/util/random.cpp has no reseed function at all -- its
  seed/xorIndex/xorKey statics just start at 0 and mutate as rand() gets
  called. OpenSWOS's int-splitting Reseed() policy is its own invention.
  Re-tagged PORT_EXTENSION in swos_rng.h, with a note that seed=0 is the one
  input verified to reproduce the original's true start state exactly.
- Added RNG golden vectors (seed 0/42/0x12345678, both streams) computed
  independently in Python against random.cpp's algorithm + table, not by
  calling this repo's own C -- a same-seed-twice determinism test alone
  can't catch a wrong-but-self-consistent algorithm.
- Added flag boundary tests: signed overflow (add16 0x7fff+1, sub16
  0x8000-1) and unsigned carry-out (add16 0xffff+1, sub16 0-1).
- Added assert()-based bounds checking (no-op under NDEBUG) to every Read/
  Write/View call.
- Added .gitattributes (force LF) so regenerating the two generated headers
  doesn't show as line-ending-only diffs on this machine's autocrlf setup.
- Fixed swos_rng.h's stale "random.cpp not present in this workspace"
  comment -- it's at ../swos-port/src/util/random.cpp, and now cited there.

make test: 26/26 checks pass.

+++ include/swos_memory.h
@@ -24,6 +24,10 @@
 // segment 0x00000..0x4F7FF, sprite pool 0x4F800..0x5FFFF).
 #define SWOS_MEM_SIZE 0x60000
 
+// All Read*/Write* helpers and swosMemoryView() bounds-check addr via
+// assert() (a no-op under NDEBUG, so this costs nothing in a release/DS
+// build; the desktop test build keeps it active).
+
 // ---- Read helpers -----------------------------------------------------
 uint8_t swosReadByte(int addr);
 uint16_t swosReadWord(int addr);          // unsigned word
@@ -32,9 +36,14 @@ uint32_t swosReadDword(int addr);
 int32_t swosReadSignedDword(int addr);
 
 // ---- Write helpers ----------------------------------------------------
+// Word/dword take unsigned value types (not the C# port's plain `int`):
+// converting a value like 0xDEADBEEF to a 32-bit *signed* int, or
+// right-shifting a negative int, are both implementation-defined in C. To
+// write a signed value, cast explicitly at the call site:
+// swosWriteDword(addr, (uint32_t)signedValue).
 void swosWriteByte(int addr, int value);
-void swosWriteWord(int addr, int value);
-void swosWriteDword(int addr, int value);
+void swosWriteWord(int addr, uint16_t value);
+void swosWriteDword(int addr, uint32_t value);
 
 // ---- Lifecycle ----------------------------------------------------------
 // Zeroes the backing buffer only. Placeholder for the real swosMemoryInit()
